refactor: ts migration, add .env.test
